+++ main.go
@@ -18,6 +18,7 @@ import (
 	"log"
 	"notifications/core"
 	"notifications/core/model"
+	"notifications/driven/airship"
 	corebb "notifications/driven/core"
 	"notifications/driven/firebase"
 	"notifications/driven/mailer"
@@ -27,9 +28,9 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/golang-jwt/jwt"
-	"github.com/rokwire/core-auth-library-go/v2/authservice"
-	"github.com/rokwire/core-auth-library-go/v2/sigauth"
+	"github.com/rokwire/core-auth-library-go/v3/authservice"
+	"github.com/rokwire/core-auth-library-go/v3/keys"
+	"github.com/rokwire/core-auth-library-go/v3/sigauth"
 	"github.com/rokwire/logging-library-go/v2/logs"
 )
 
@@ -65,25 +66,31 @@ func main() {
 	storageAdapter := storage.NewStorageAdapter(mongoDBAuth, mongoDBName, mongoTimeout, mtOrgID, mtAppID, logger)
 	err := storageAdapter.Start()
 	if err != nil {
-		log.Fatal("Cannot start the mongoDB adapter - " + err.Error())
+		logger.Fatal("Cannot start the mongoDB adapter - " + err.Error())
 	}
 
 	// firebase adapter
 	firebaseConfs, err := storageAdapter.LoadFirebaseConfigurations()
+	_, err = storageAdapter.LoadFirebaseConfigurations()
 	if err != nil {
-		log.Fatal("Error loading the firebase confogirations from the storage - " + err.Error())
+		logger.Fatal("Error loading the firebase configurations from the storage - " + err.Error())
 	}
 	firebaseAdapter := firebase.NewFirebaseAdapter()
 	err = firebaseAdapter.Start(firebaseConfs)
 	if err != nil {
-		log.Fatal("Cannot start the Firebase adapter - " + err.Error())
+		logger.Warn("Cannot start the Firebase adapter - " + err.Error())
 	}
 
-	smtpHost := getEnvKey("SMTP_HOST", true)
-	smtpPort := getEnvKey("SMTP_PORT", true)
-	smtpUser := getEnvKey("SMTP_USER", true)
-	smtpPassword := getEnvKey("SMTP_PASSWORD", true)
-	smtpFrom := getEnvKey("SMTP_EMAIL_FROM", true)
+	//airship adapter
+	airshipHost := getEnvKey("NOTIFICATIONS_AIRSHIP_HOST", false)
+	airshipBearerToken := getEnvKey("NOTIFICATIONS_AIRSHIP_BEARER_TOKEN", false)
+	airshipAdapter := airship.NewAirshipAdapter(airshipHost, airshipBearerToken)
+
+	smtpHost := getEnvKey("SMTP_HOST", false)
+	smtpPort := getEnvKey("SMTP_PORT", false)
+	smtpUser := getEnvKey("SMTP_USER", false)
+	smtpPassword := getEnvKey("SMTP_PASSWORD", false)
+	smtpFrom := getEnvKey("SMTP_EMAIL_FROM", false)
 	smtpPortNum, _ := strconv.Atoi(smtpPort)
 	mailAdapter := mailer.NewMailerAdapter(smtpHost, smtpPortNum, smtpUser, smtpPassword, smtpFrom)
 
@@ -102,35 +109,38 @@ func main() {
 
 	serviceRegLoader, err := authservice.NewRemoteServiceRegLoader(&authService, []string{"auth"})
 	if err != nil {
-		log.Fatalf("Error initializing remote service registration loader: %v", err)
+		logger.Fatalf("Error initializing remote service registration loader: %v", err)
 	}
 
-	serviceRegManager, err := authservice.NewServiceRegManager(&authService, serviceRegLoader)
+	serviceRegManager, err := authservice.NewServiceRegManager(&authService, serviceRegLoader, !strings.HasPrefix(host, "http://localhost"))
 	if err != nil {
-		log.Fatalf("Error initializing service registration manager: %v", err)
+		logger.Fatalf("Error initializing service registration manager: %v", err)
 	}
 
 	//core adapter
 	serviceAccountID := getEnvKey("NOTIFICATIONS_SERVICE_ACCOUNT_ID", false)
-	privKeyRaw := getEnvKey("NOTIFICATIONS_PRIV_KEY", true)
-	privKeyRaw = strings.ReplaceAll(privKeyRaw, "\\n", "\n")
-	privKey, err := jwt.ParseRSAPrivateKeyFromPEM([]byte(privKeyRaw))
-	if err != nil {
-		log.Fatalf("Error parsing priv key: %v", err)
-	}
-	signatureAuth, err := sigauth.NewSignatureAuth(privKey, serviceRegManager, false)
-	if err != nil {
-		log.Fatalf("Error initializing signature auth: %v", err)
-	}
+	privKeyRaw := getEnvKey("NOTIFICATIONS_PRIV_KEY", false)
+	var serviceAccountManager *authservice.ServiceAccountManager
+	if privKeyRaw != "" {
+		privKeyRaw = strings.ReplaceAll(privKeyRaw, "\\n", "\n")
+		privKey, err := keys.NewPrivKey(keys.RS256, privKeyRaw)
+		if err != nil {
+			log.Fatalf("Failed to parse auth priv key: %v", err)
+		}
+		signatureAuth, err := sigauth.NewSignatureAuth(privKey, serviceRegManager, false, false)
+		if err != nil {
+			log.Fatalf("Error initializing signature auth: %v", err)
+		}
 
-	serviceAccountLoader, err := authservice.NewRemoteServiceAccountLoader(&authService, serviceAccountID, signatureAuth)
-	if err != nil {
-		log.Fatalf("Error initializing remote service account loader: %v", err)
-	}
+		serviceAccountLoader, err := authservice.NewRemoteServiceAccountLoader(&authService, serviceAccountID, signatureAuth)
+		if err != nil {
+			log.Fatalf("Error initializing remote service account loader: %v", err)
+		}
 
-	serviceAccountManager, err := authservice.NewServiceAccountManager(&authService, serviceAccountLoader)
-	if err != nil {
-		log.Fatalf("Error initializing service account manager: %v", err)
+		serviceAccountManager, err = authservice.NewServiceAccountManager(&authService, serviceAccountLoader)
+		if err != nil {
+			log.Fatalf("Error initializing service account manager: %v", err)
+		}
 	}
 
 	coreAdapter := corebb.NewCoreAdapter(coreBBHost, serviceAccountManager)
@@ -142,7 +152,7 @@ func main() {
 	}
 
 	// application
-	application := core.NewApplication(Version, Build, storageAdapter, firebaseAdapter, mailAdapter, logger, coreAdapter)
+	application := core.NewApplication(Version, Build, storageAdapter, firebaseAdapter, mailAdapter, logger, coreAdapter, airshipAdapter)
 	application.Start()
 
 	webAdapter := driver.NewWebAdapter(host, port, application, config, serviceRegManager, logger)