+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add in Airship Push Notifications [#173](https://github.com/rokwire/notifications-building-block/issues/173)
+
 ## [1.19.0] - 2023-10-26
 ## [1.18.0] - 2023-09-20
 ### Changed